fix: Fix PostgreSQL backup with LVM snapshots and cleanup warnings
Database Strategy Registration:
- Add alias support to BackupService::registerDatabaseStrategy()
- Register MysqlBackupStrategy under both 'mysql' and 'mariadb'
- Register PostgresBackupStrategy under both 'postgres' and 'postgresql'
- Fixes "No backup strategy registered for type: postgresql" error

LVM Volume Detection:
- Fix LVM name parsing in CapabilitiesDetectionHandler
- Use actual names from 'lvs' output instead of regex parsing
- Prevents invalid LV names like "-vg-ubuntu--lv" from /dev/mapper parsing
- Correctly extracts VG and LV names for volumes with hyphens

Snapshot Size Format:
- Add 'recommended_lvm' field without space for LVM commands (e.g., "1G" not "1 GB")
- Update BackupWizardController to use LVM-compatible format
- Fixes "lvcreate: Logical volume name is invalid" errors

Database Info Creation:
- Add repo_id parameter to DatabaseInfoRepository::create()
- Include repo_id and pg_svg_path in INSERT to satisfy NOT NULL constraints
- Remove redundant updateRepositoryId() call after create
- Fixes "Field 'repo_id' doesn't have a default value" error

PHP Warnings Cleanup:
- Fix float-to-int implicit conversion in BackupController duration formatting
- Remove undefined createdAt/updatedAt properties from ServerController
- Cast floor() results explicitly to int for sprintf()

Testing:
- ✅ First successful PostgreSQL LVM snapshot backup
- ✅ Archive created: postgresql_2025-11-15_15-26-23
- ✅ Stats: 83MB original → 15MB deduplicated (81% space saved)
- ✅ Snapshot cleanup working correctly

+++ src/Api/Controller/BackupController.php
@@ -275,9 +275,9 @@ class BackupController extends BaseController
         
         // Format duration
         $duration = (float)($row['dur'] ?? 0);
-        $hours = floor($duration / 3600);
-        $minutes = floor(($duration % 3600) / 60);
-        $seconds = $duration % 60;
+        $hours = (int)floor($duration / 3600);
+        $minutes = (int)floor(($duration % 3600) / 60);
+        $seconds = (int)($duration % 60);
         $durationFormatted = sprintf('%02d:%02d:%02d', $hours, $minutes, $seconds);
 
         return [
